# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: ThreatTrace AI — Core REST API
  description: Production-grade REST API for ThreatTrace AI Threat Intelligence & Risk Analysis Platform.
  version: 1.0.0
  contact:
    name: ThreatTrace Security Architecture Team
servers:
  - url: http://localhost:8080/api/v1
    description: Local Core Backend
paths:
  /auth/login:
    post:
      summary: User Authentication
      description: Authenticates user credentials and returns JWT access and refresh tokens.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username:
                  type: string
                  example: analyst@threattrace.io
                password:
                  type: string
                  format: password
                  example: SecurePass123!
      responses:
        '200':
          description: Successful authentication
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /auth/register:
    post:
      summary: Register New User
      description: Creates a new user account with default VIEWER/ANALYST role.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: User registered successfully
        '400':
          description: Validation error

  /auth/me:
    get:
      summary: Current Authenticated User Profile
      security:
        - BearerAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'

  /threats:
    get:
      summary: Search and Filter Threats
      security:
        - BearerAuth: []
      parameters:
        - name: query
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
        - name: severity
          in: query
          schema:
            type: string
        - name: minRisk
          in: query
          schema:
            type: integer
        - name: maxRisk
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
        - name: sourceType
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedThreatResponse'
    post:
      summary: Ingest Suspicious Threat Report
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ThreatIngestRequest'
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ThreatDetailResponse'

  /threats/{id}:
    get:
      summary: Retrieve Threat Details & AI Analysis
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ThreatDetailResponse'
        '404':
          description: Threat not found

  /threats/{id}/reanalyze:
    post:
      summary: Trigger Re-analysis of Threat
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ThreatDetailResponse'

  /dashboard/kpi:
    get:
      summary: Retrieve Top-Level SOC Metrics
      security:
        - BearerAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DashboardKPIResponse'

  /dashboard/timeline:
    get:
      summary: Retrieve Threat Activity Timeline
      security:
        - BearerAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TimelineEvent'

  /investigations:
    get:
      summary: List Investigations
      security:
        - BearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: priority
          in: query
          schema:
            type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/InvestigationResponse'
    post:
      summary: Escalate Threat to Investigation
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateInvestigationRequest'
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvestigationResponse'

  /alerts:
    get:
      summary: List Active Alerts
      security:
        - BearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AlertResponse'

  /alerts/{id}/acknowledge:
    post:
      summary: Acknowledge an Alert
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlertResponse'

  /audit:
    get:
      summary: Query Security Audit Trail
      security:
        - BearerAuth: []
      parameters:
        - name: action
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 50
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedAuditResponse'

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    AuthResponse:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        tokenType:
          type: string
          example: Bearer
        expiresIn:
          type: integer
          example: 900
        user:
          $ref: '#/components/schemas/UserProfile'

    UserProfile:
      type: object
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        email:
          type: string
        fullName:
          type: string
        roles:
          type: array
          items:
            type: string

    RegisterRequest:
      type: object
      required: [username, email, password, fullName]
      properties:
        username:
          type: string
          minLength: 3
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        fullName:
          type: string

    ThreatIngestRequest:
      type: object
      required: [sourceType, rawContent]
      properties:
        sourceType:
          type: string
          enum: [EMAIL, WEB_REPORT, SOCIAL_MEDIA, MESSAGE, USER_REPORT, API, FILE_UPLOAD]
        sourceReference:
          type: string
        rawContent:
          type: string
          minLength: 5
          maxLength: 100000
        locationName:
          type: string
        latitude:
          type: number
        longitude:
          type: number

    ThreatDetailResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        threatReference:
          type: string
          example: THR-2026-000184
        sourceType:
          type: string
        sourceReference:
          type: string
        rawContent:
          type: string
        status:
          type: string
        locationName:
          type: string
        latitude:
          type: number
        longitude:
          type: number
        geoVerified:
          type: boolean
        createdAt:
          type: string
          format: date-time
        analysis:
          $ref: '#/components/schemas/ThreatAnalysisDTO'

    ThreatAnalysisDTO:
      type: object
      properties:
        predictedCategory:
          type: string
        confidence:
          type: number
        riskScore:
          type: integer
          example: 82
        severity:
          type: string
          enum: [LOW, MEDIUM, HIGH, CRITICAL]
        sentimentLabel:
          type: string
        sentimentScore:
          type: number
        urgencyScore:
          type: number
        explanationSummary:
          type: string
        modelName:
          type: string
        modelVersion:
          type: string
        humanReviewRequired:
          type: boolean
        signalBreakdown:
          type: array
          items:
            type: object
            properties:
              signal:
                type: string
              points:
                type: integer
              detail:
                type: string
        entities:
          type: array
          items:
            $ref: '#/components/schemas/ThreatEntityDTO'
        indicators:
          type: array
          items:
            $ref: '#/components/schemas/ThreatIndicatorDTO'

    ThreatEntityDTO:
      type: object
      properties:
        entityType:
          type: string
        entityValue:
          type: string
        confidence:
          type: number
        startOffset:
          type: integer
        endOffset:
          type: integer
        sourceSpan:
          type: string

    ThreatIndicatorDTO:
      type: object
      properties:
        indicatorType:
          type: string
        label:
          type: string
        weight:
          type: number
        evidenceSnippet:
          type: string

    DashboardKPIResponse:
      type: object
      properties:
        totalThreats:
          type: integer
        highRiskThreats:
          type: integer
        activeInvestigations:
          type: integer
        openAlerts:
          type: integer
        requiringReview:
          type: integer
        averageRiskScore:
          type: number

    TimelineEvent:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        threatId:
          type: string
        threatReference:
          type: string
        source:
          type: string
        category:
          type: string
        severity:
          type: string
        riskScore:
          type: integer
        status:
          type: string
        location:
          type: string

    InvestigationResponse:
      type: object
      properties:
        id:
          type: string
        investigationReference:
          type: string
        threatId:
          type: string
        threatReference:
          type: string
        priority:
          type: string
        status:
          type: string
        title:
          type: string
        summary:
          type: string
        assignedAnalyst:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string

    AlertResponse:
      type: object
      properties:
        id:
          type: string
        alertReference:
          type: string
        threatId:
          type: string
        threatReference:
          type: string
        alertType:
          type: string
        severity:
          type: string
        status:
          type: string
        createdAt:
          type: string

    PagedThreatResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/ThreatDetailResponse'
        totalElements:
          type: integer
        totalPages:
          type: integer
        size:
          type: integer
        number:
          type: integer

    PagedAuditResponse:
      type: object
      properties:
        content:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              actor:
                type: string
              action:
                type: string
              resourceType:
                type: string
              resourceId:
                type: string
              ipAddress:
                type: string
              timestamp:
                type: string
        totalElements:
          type: integer
        totalPages:
          type: integer

    ErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
          example: 400
        error:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
        path:
          type: string
        requestId:
          type: string